+++ tests/test_evaluation.py
@@ -25,7 +25,7 @@ class TestFindClosestMatchesFunction(unittest.TestCase):
     def test_types(self):
         matches = find_closest_matches([], [])
         self.assertIsInstance(matches, np.ndarray)
-        self.assertEqual(matches.dtype, np.int)
+        self.assertEqual(matches.dtype, int)
         self.assertIsInstance(find_closest_matches([], []), np.ndarray)
 
     def test_value(self):
@@ -46,7 +46,7 @@ class TestCalcErrorsFunction(unittest.TestCase):
     def test_types(self):
         errors = calc_errors(DETECTIONS, ANNOTATIONS)
         self.assertIsInstance(errors, np.ndarray)
-        self.assertEqual(errors.dtype, np.float)
+        self.assertEqual(errors.dtype, float)
 
     def test_values(self):
         # empty sequences
@@ -70,7 +70,7 @@ class TestCalcAbsoluteErrorsFunction(unittest.TestCase):
     def test_types(self):
         errors = calc_absolute_errors(DETECTIONS, ANNOTATIONS)
         self.assertIsInstance(errors, np.ndarray)
-        self.assertEqual(errors.dtype, np.float)
+        self.assertEqual(errors.dtype, float)
 
     def test_values(self):
         # empty sequences
@@ -270,13 +270,13 @@ class TestEvaluationClass(unittest.TestCase):
                           'fmeasure', 'accuracy'])
         # test with other values
         e = Evaluation(tp=[1, 2, 3.0], fp=[1.5], fn=[0, 3.1])
-        tp = np.asarray([1, 2, 3], dtype=np.float)
+        tp = np.asarray([1, 2, 3], dtype=float)
         self.assertTrue(np.allclose(e.tp, tp))
-        fp = np.asarray([1.5], dtype=np.float)
+        fp = np.asarray([1.5], dtype=float)
         self.assertTrue(np.allclose(e.fp, fp))
-        tn = np.asarray([], dtype=np.float)
+        tn = np.asarray([], dtype=float)
         self.assertTrue(np.allclose(e.tn, tn))
-        fn = np.asarray([0, 3.1], dtype=np.float)
+        fn = np.asarray([0, 3.1], dtype=float)
         self.assertTrue(np.allclose(e.fn, fn))
         self.assertEqual(e.num_tp, 3)
         self.assertEqual(e.num_fp, 1)
